+++ .github/workflows/fuzzers.yml
@@ -14,7 +14,7 @@ jobs:
   build:    
     runs-on: ubuntu-latest
     env:
-      allfuzzers: parser dump
+      allfuzzers: parser dump dump_raw_tape print_json
       artifactsprefix: -artifact_prefix=fuzzfailure/
     steps:
     - name: Install packages necessary for building
